feat: Added delete option functionnal

+++ model/article.php
@@ -41,15 +41,19 @@ class Article
 
     public function editArticle(string $title, ?array $tags, string $image, string $paragraph): bool
     {
+        global $connectedUser;
         try{
-            if($image) move_uploaded_file($_FILES['image']['tmp_name'], "images/" . $image);
-            else $image = "unavailable.png";
-            
-            Database::request("DELETE FROM `article_tag` WHERE article_id= ?;", [$this->articleId]);
-            if($tags) foreach($tags as $tag) Database::request("INSERT INTO article_tag (article_id, tag_id) VALUES (?, ?)", [$this->articleId, Tag::getTagId($tag)]);
+            if($connectedUser->id === $this->idClient) {
+                if($image) move_uploaded_file($_FILES['image']['tmp_name'], "images/" . $image);
+                else $image = "unavailable.png";
+                
+                Database::request("DELETE FROM `article_tag` WHERE article_id= ?;", [$this->articleId]);
+                if($tags) foreach($tags as $tag) Database::request("INSERT INTO article_tag (article_id, tag_id) VALUES (?, ?)", [$this->articleId, Tag::getTagId($tag)]);
 
-            Database::request("UPDATE articles SET articleTitle= ?, articleImage= ?, articleParagraph= ? WHERE articleId= ?;", [$title, $image, $paragraph, $this->articleId]);
-            return true;
+                Database::request("UPDATE articles SET articleTitle= ?, articleImage= ?, articleParagraph= ?, approuve= ? WHERE articleId= ?;", [$title, $image, $paragraph, 0, $this->articleId]);
+                return true;
+            }
+            return false;
 
         }catch (Exception $e) {return false;}
     }
@@ -63,9 +67,14 @@ class Article
 
     public function removeArticle(): bool
     {
+        global $connectedUser;
         try{
-            Database::request("DELETE FROM articles WHERE articleId= ?;", [$this->articleId]);
-            return true;
+            if($connectedUser->id === $this->idClient) {
+                Database::request("DELETE FROM articles WHERE articleId= ?;", [$this->articleId]);
+                return true;
+            }
+            
+            return false;
 
         }catch (Exception $e) {return false;}
     }
